+++ example/handlers.go
@@ -267,6 +267,7 @@ func (a app) login(secretKey string) http.HandlerFunc {
 
 // handleFileServer handler showcases the use of:
 // - middleware.ClientIP
+// - middleware.ClientFingerPrint
 // - logging
 func (a app) handleFileServer() http.HandlerFunc {
 	// Do NOT let `http.FileServer` be able to serve your root directory.
@@ -284,7 +285,7 @@ func (a app) handleFileServer() http.HandlerFunc {
 	realHandler := http.StripPrefix("/staticAssets/", fs).ServeHTTP
 	return func(w http.ResponseWriter, r *http.Request) {
 		reqL := log.WithID(r.Context(), a.l)
-		reqL.Info("handleFileServer", "clientIP", middleware.ClientIP(r))
+		reqL.Info("handleFileServer", "clientIP", middleware.ClientIP(r), "clientFingerPrint", middleware.ClientFingerPrint(r))
 
 		slog.NewLogLogger(reqL.Handler(), log.LevelImmediate).
 			Println("this is now a Go standard library logger")