+++ internal/octx/octx.go
@@ -1,8 +1,16 @@
 // Package octx houses context keys used by multiple ong packages.
 package octx
 
-type logContextKeyType string
+type (
+	logContextKeyType  string
+	fingerPrintKeyType string
+)
 
-// LogCtxKey is the name of the context key used to store the logID.
-// It is used primarily by `ong/log`, `ong/client` and `ong/middleware` packages
-const LogCtxKey = logContextKeyType("Ong-logID")
+const (
+	// LogCtxKey is the name of the context key used to store the logID.
+	// It is used primarily by `ong/log`, `ong/client` and `ong/middleware` packages
+	LogCtxKey = logContextKeyType("Ong-logID")
+
+	// FingerPrintCtxKey is the name of the context key used to store the TLS fingerprint.
+	FingerPrintCtxKey = fingerPrintKeyType("fingerPrintKeyType")
+)