+++ server/cert_test.go
@@ -14,11 +14,6 @@ import (
 func TestCreateDevCertKey(t *testing.T) {
 	t.Parallel()
 
-	if os.Getenv("GITHUB_ACTIONS") != "" {
-		// CreateDevCertKey() fails in github actions with error: `panic: open /home/runner/ong/rootCA_key.pem: permission denied`
-		return
-	}
-
 	certPath, keyPath := certKeyPaths()
 
 	t.Run("success", func(t *testing.T) {