+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 Most recent version is listed first.  
 
 
+## v0.0.45
+- Run all tests in CI: https://github.com/komuw/ong/pull/248
+
 ## v0.0.44
 - Organise imports: https://github.com/komuw/ong/pull/245
 - Create an internal/octx that houses context keys used by multiple ong packages: https://github.com/komuw/ong/pull/246